requirements.txt
gql[requests]==3.4.0
pandas==1.4.3
python-dotenv==0.20.0
tweepy==4.10.0